feat experimets

+++ ssk_modules/string_kernel.pyx
@@ -27,8 +27,8 @@ def ssk_array(array.array s_, array.array t_, int n, float lbda, accum=False):
 
     lens, lent = len(s), len(t)
     #k_prim = (-1)*np.ones( (n+1, lens, lent) )
-    cdef np.ndarray[float, ndim=3] \
-        k_prim = np.zeros((n, lens, lent), dtype=float)
+    cdef np.ndarray[np.float64_t, ndim=3] \
+        k_prim = np.zeros((n, lens, lent), dtype=np.float)
 
     k_prim[0,:,:] = 1
 
@@ -61,7 +61,7 @@ def string_kernel(xs, ys, n, lbda):
 
     cdef int lenxs, lenys
     cdef int i, j
-    cdef np.ndarray[float, ndim=2] mat, mat_xs, mat_ys
+    cdef np.ndarray[np.float64_t, ndim=2] mat, mat_xs, mat_ys
     lenxs, lenys = xs.shape[0], ys.shape[0]
 
     mat = np.zeros((lenxs, lenys))
